agent/requirements.txt
web3==6.15.0
requests==2.31.0
anthropic==0.75.0
python-dotenv==1.0.0
